Fixes for new firmware compatibility (#41)
Addresses issues #38 & #39

Context for #38:
Old protocol implementation sends an empty buffer after sending a header indicating a scan saw 0 devices. New implementation doesn't, for obvious reasons. This causes the hub to crash with the new implementation, because it is expecting an empty buffer and receives a (nonempty) scan header instead.

Fix for #38:
Write the empty scan in the scan header processing section, instead of in the scan data processing section. Then, handle the next received packet as a scan header. In the header processing section, before we clear the tempScan data, check to see if the last scan had no devices and if we received an empty buffer. If both of those are true, then we skip the packet and expect a scan header next. If either of those are false, we handle the packet as normal.

Context for #39:
In the new implementation, the clock starts when the badges are turned on, instead of when they are synchronized. Since we checked the timestamp to see if the badges were synced, this obviously doesn't work with the new firmware.

Fix for #39:
Use the clockSet value (which is already being passed in the status header from the badges) to check if the badges were synchronized.